[Feat/#77] ConversationView에 스크롤 컨테이너 적용 + 코드 정리
