+++ talklat/talklat/Sources/Stores/ConversationViewStore.swift
@@ -22,6 +22,14 @@ final class ConversationViewStore: ObservableObject {
         var hasChevronButtonTapped: Bool = false
         var historyItems: [HistoryItem] = []
         var historyItem: HistoryItem?
+        
+        // scroll container related - TODO: ScrollStore 분리?
+        var historyScrollViewHeight: CGFloat = CGFloat(0)
+        var historyScrollOffset: CGPoint = CGPoint(x: -0.0, y: 940.0)
+        var deviceHeight: CGFloat = 0
+        var topInset: CGFloat = 0
+        var bottomInset: CGFloat = 0
+        var isTopViewShown: Bool = false
     }
     
     @Published private var viewState: ConversationState
@@ -59,6 +67,26 @@ final class ConversationViewStore: ObservableObject {
                 }
             }
         )
+    }
+    
+    public func bindingHistoryScrollOffset() -> Binding<CGPoint> {
+        Binding(
+            get: { self(\.historyScrollOffset) },
+            set: { _ in self(\.historyScrollOffset) }
+        )
+    }
+    
+    public func onScrollContainerAppear(
+        geo: GeometryProxy,
+        insets: UIEdgeInsets
+    ) {
+        reduce(\.deviceHeight, into: geo.size.height)
+        reduce(\.topInset, into: insets.top)
+        reduce(\.bottomInset, into: insets.bottom)
+    }
+    
+    public func onHistoryViewAppear(geo: GeometryProxy) {
+        reduce(\.historyScrollViewHeight, into: geo.size.height)
     }
     
     public func onWritingViewAppear() {
@@ -116,6 +144,10 @@ final class ConversationViewStore: ObservableObject {
              ? false
              : true
         )
+    }
+    
+    public func onScrollOffsetChanged(_ value: Bool) {
+        reduce(\.isTopViewShown, into: value)
     }
     
     public func onGuideCancelButtonTapped() {
