+++ talklat/talklat/Sources/Views/TKHistoryView.swift
@@ -7,74 +7,40 @@
 
 import SwiftUI
 
-import SwiftUI
-
 struct TKHistoryView: View {
     @Environment(\.colorScheme)
     var colorScheme
-    @ObservedObject var appViewStore: AppViewStore
+    
+    @ObservedObject var store: ConversationViewStore
     
     var body: some View {
         ScrollViewReader { proxy in
+            Rectangle()
+                .fill(Color.clear)
+                .frame(height: 30)
+            
             OffsetObservingScrollView(
-                offset: $appViewStore.historyScrollOffset
+                offset: store.bindingHistoryScrollOffset()
             ) {
                 // TODO: - ForEach의 data 아규먼트 수정
                 // TODO: - 각 Color 값을 디자인 시스템 값으로 추후 수정
                 ForEach(
-                    appViewStore.historyItems,
+                    store(\.historyItems),
                     id: \.id
                 ) { item in
                     switch item.type {
                     case .question:
-                        Text(item.text)
-                            .font(.subheadline)
-                            .foregroundColor(.gray700)
-                            .padding(.horizontal, 16)
-                            .padding(.vertical, 8)
-                            .background {
-                                RoundedRectangle(cornerRadius: 12)
-                                    .fill(Color.gray200)
-                            }
-                            .frame(
-                                maxWidth: .infinity,
-                                alignment: .trailing
-                            )
-                            .padding(.trailing, 24)
-                            .padding(.leading, 68)
-                            .padding(.top, 32)
+                        questionTextBuilder(item)
                         
                     case .answer:
-                        VStack(alignment: .leading) {
-                            Image(systemName: "waveform.circle.fill")
-                                .resizable()
-                                .frame(width: 24, height: 24)
-                                .foregroundColor(Color.gray700)
-                                .padding(.leading, 4)
-                            
-                            Text(item.text)
-                                .font(.headline)
-                                .foregroundColor(.gray700)
-                                .padding(.horizontal, 16)
-                                .padding(.vertical, 8)
-                                .background {
-                                    RoundedRectangle(cornerRadius: 12)
-                                        .fill(Color(.white))
-                                }
-                        }
-                        .frame(
-                            maxWidth: .infinity,
-                            alignment: .leading
-                        )
-                        .padding(.horizontal, 24)
-                        .id(appViewStore.checkIfLastItem(item))
+                        answerTextBuilder(item)
                     }
                 }
                 .padding(.top, 10)
                 .background(
                     GeometryReader { geometry in
                         Color.clear.onAppear {
-                            appViewStore.historyScrollViewHeight = geometry.size.height
+                            store.onHistoryViewAppear(geo: geometry)
                         }
                     }
                 )
@@ -85,43 +51,27 @@ struct TKHistoryView: View {
                     anchor: .bottom
                 )
             }
-            // TODO: - deviceTopSafeAreaInset 값으로 변경
-            .padding(.top, 100)
-            .background { Color(.systemGray6 )}
-            .safeAreaInset(edge: .bottom) {
-                VStack {
-                    swipeGuideMessage(type: .swipeToBottom)
-                        .offset(appViewStore.messageOffset)
-                        .frame(
-                            maxWidth: .infinity,
-                            maxHeight: 100,
-                            alignment: .top
-                        )
-                        .background {
-                            Rectangle()
-                                .fill(checkColorScheme())
-                                .shadow(
-                                    color: Color.gray300,
-                                    radius: 25,
-                                    y: -5
-                                )
-                        }
+            .navigationTitle(
+                store(\.isTopViewShown)
+                ? "히스토리"
+                : ""
+            )
+        }
+        .ignoresSafeArea()
+        .onChange(of: store(\.historyScrollOffset)) { offset in
+            print("---> history offset: ", offset)
+            
+            DispatchQueue.main.asyncAfter(
+                deadline: .now() + 0.3
+            ) {
+                if offset.y > 870,
+                   offset.y < 920,
+                   store.isChevronButtonDisplayable {
+                    withAnimation(.spring(dampingFraction: 0.7)) {
+                        store.onScrollOffsetChanged(false)
+                    }
                 }
             }
-            .ignoresSafeArea(edges: .top)
-            .background { Color.gray100 }
-            .navigationTitle(
-                appViewStore.isHistoryViewShown ? "히스토리" : ""
-            )
-            .navigationBarTitleDisplayMode(.inline)
-            .toolbarBackground(
-                Color(.systemGray6),
-                for: .navigationBar
-            )
-            .toolbarBackground(
-                .visible,
-                for: .navigationBar
-            )
         }
     }
     
@@ -134,14 +84,50 @@ struct TKHistoryView: View {
     }
 }
 
-struct TKHistoryView_Previews: PreviewProvider {
-    static var previews: some View {
-        TKHistoryView(appViewStore: .makePreviewStore(condition: { store in
-            store.historyItems.append(.init(id: .init(), text: "대답1", type: .answer))
-            store.historyItems.append(.init(id: .init(), text: "질문1", type: .question))
-            store.historyItems.append(.init(id: .init(), text: "일이삼사오육칠팔구일이삼사오육칠팔구일이삼사오육칠팔구일이삼사오육칠팔구일이삼사오육칠팔구일이삼사오육칠팔구일이삼사오육칠팔구일이삼사오육칠팔구일이삼사오육칠팔구일이삼사오육칠팔구일이삼사오육칠팔구일이삼사오육칠팔구1", type: .answer))
-            store.historyItems.append(.init(id: .init(), text: "일이삼사오육칠팔구일이삼사오육칠팔구일이삼사오육칠팔구일이삼사오육칠팔구일이삼사오육칠팔구일이삼사오육칠팔구일이삼사오육칠팔구일이삼사오육칠팔구일이삼사오육칠팔구일이삼사오육칠팔구일이삼사오육칠팔구일이삼사오육칠팔구1", type: .question))
-            store.historyItems.append(.init(id: .init(), text: "일이삼사오육칠팔구일이삼사오육칠팔구일이삼사오육칠팔구일이삼사오육칠팔구일이삼사오육칠팔구일이삼사오육칠팔구일이삼사오육칠팔구일이삼사오육칠팔구일이삼사오육칠팔구일이삼사오육칠팔구일이삼사오육칠팔구일이삼사오육칠팔구1", type: .answer))
-        }))
+extension TKHistoryView {
+    private func questionTextBuilder(_ item: HistoryItem) -> some View {
+        Text(item.text)
+            .font(.subheadline)
+            .foregroundColor(.gray700)
+            .padding(.horizontal, 16)
+            .padding(.vertical, 8)
+            .background {
+                RoundedRectangle(cornerRadius: 12)
+                    .fill(Color.gray200)
+            }
+            .frame(
+                maxWidth: .infinity,
+                alignment: .trailing
+            )
+            .padding(.trailing, 24)
+            .padding(.leading, 68)
+            .padding(.top, 32)
+    }
+    
+    
+    private func answerTextBuilder(_ item: HistoryItem) -> some View {
+        VStack(alignment: .leading) {
+            Image(systemName: "waveform.circle.fill")
+                .resizable()
+                .frame(width: 24, height: 24)
+                .foregroundColor(Color.gray700)
+                .padding(.leading, 4)
+            
+            Text(item.text)
+                .font(.headline)
+                .foregroundColor(.gray700)
+                .padding(.horizontal, 16)
+                .padding(.vertical, 8)
+                .background {
+                    RoundedRectangle(cornerRadius: 12)
+                        .fill(Color(.white))
+                }
+        }
+        .frame(
+            maxWidth: .infinity,
+            alignment: .leading
+        )
+        .padding(.horizontal, 24)
+//        .id(store(\.historyScrollOffset).checkIfLastItem(item))
     }
 }
